+++ CHANGELOG.md
@@ -38,6 +38,15 @@
     there is more than one object with the same name in the recycle bin.
     Now it uses the object that was changed last using the property
     `whenChanged` ([issue #271](https://github.com/PowerShell/xActiveDirectory/issues/271)).
+  - BREAKING CHANGE: Previously a computer account was always set to enabled
+    regardless of `Enabled` parameter was used in a configuration. Now if
+    the `Enabled` parameter is left out of the configuration, the
+    computer account will not be evaluated if it enabled. So if a computer is
+    disabled, the resource will not enabled it unless `Enabled` parameter is
+    set to `$true` in the configuration.
+  - A computer account can now be created disabled by setting the parameter
+    `CreateDisabled` to `$true`. Default a computer account will be created
+    as enabled.
 - Changes to xADGroup
   - Restoring a group from the recycle bin no longer fails if there is
     more than one object with the same name in the recycle bin. Now it